+++ packages/vulcan-ui-material/lib/components/forms/base-controls/FormSelect.jsx
@@ -3,8 +3,8 @@ import React from 'react';
 import createReactClass from 'create-react-class';
 import PropTypes from 'prop-types';
 import ComponentMixin from './mixins/component';
-import MuiFormControl from './MuiFormControl';
-import MuiFormHelper from './MuiFormHelper';
+import FormControlLayout from './FormControlLayout';
+import FormHelper from './FormHelper';
 import Select from '@material-ui/core/Select';
 import Input from '@material-ui/core/Input';
 import MenuItem from '@material-ui/core/MenuItem';
@@ -14,10 +14,10 @@ import StartAdornment, { hideStartAdornment } from './StartAdornment';
 import EndAdornment from './EndAdornment';
 import _isArray from 'lodash/isArray';
 import classNames from 'classnames';
-import { styles } from './MuiSuggest';
+import { styles } from './FormSuggest';
 
 
-const MuiSelect = createReactClass({
+const FormSelect = createReactClass({
 
   element: null,
 
@@ -85,10 +85,10 @@ const MuiSelect = createReactClass({
     }
 
     return (
-      <MuiFormControl{...this.getFormControlProperties()} htmlFor={this.getId()}>
+      <FormControlLayout{...this.getFormControlProperties()} htmlFor={this.getId()}>
         {this.renderElement()}
-        <MuiFormHelper {...this.getFormHelperProperties()}/>
-      </MuiFormControl>
+        <FormHelper {...this.getFormHelperProperties()}/>
+      </FormControlLayout>
     );
   },
 
@@ -115,7 +115,7 @@ const MuiSelect = createReactClass({
         </MenuList>;
     };
 
-    const { options, classes } = this.props;
+    const { options = [], classes } = this.props;
 
     let groups = options.filter(function (item) {
       return item.group;
@@ -191,7 +191,7 @@ const MuiSelect = createReactClass({
                               input: classNames(classes.input, !value && classes.inputPlaceholder),
                             }}
               />}
-              classes={{ icon: classes.muiIcon }}
+              classes={{ icon: classes.selectIcon }}
       >
         {optionNodes}
       </Select>
@@ -200,4 +200,4 @@ const MuiSelect = createReactClass({
 });
 
 
-export default withStyles(styles)(MuiSelect);
+export default withStyles(styles)(FormSelect);