+++ packages/vulcan-ui-material/history.md
@@ -1,9 +1,51 @@
-1.13.2_2 / 2020-01-20
+1.16.0_2 / 2020-11-16
 =====================
 
- * MuiSuggest: Removed `selectedOption` and `inputFormatted` from the component state
- * TooltipIntl: Fixed bug: `titleValues` prop was not implemented
+ * Base controls
+   * I have renamed components in `vulcan-ui-material/lib/components/forms/base-controls/` because their names conflict with the style sheet names (used for [Global theme overrides](https://material-ui.com/customization/components/#global-theme-override)) of some core MUI components - for example `MuiInput`.
+   * These components are not registered with `registerComponent`, only exported.
+   * This will be a breaking change for anyone who has built custom components based on these base controls using import - the file and component names have to be updated.
+   * The new names are:
+     * **/forms/helpers/**
+       * MuiFormControl => FormControlLayout
+       * MuiFormHelper => FormHelper
+       * MuiRequiredIndicator => RequiredIndicator
+     * **/forms/base-controls/**
+       * MuiCheckbox => FormCheckbox
+       * MuiCheckboxGroup => FormCheckboxGroup
+       * MuiInput => FormInput
+       * MuiPicker => FormPicker
+       * MuiRadioGroup => FormRadioGroup
+       * MuiSelect => FormSelect
+       * MuiSuggest => FormSuggest
+       * MuiSwitch => FormSwitch
+       * MuiText => FormText
+ * The sample **Theme styles** page now displays sample buttons in addition to typography and color palettes
+ 
+1.16.0_1 / 2020-10-24
+=====================
 
+ * MuiSuggest
+   * Fixed how styles are applied for focused, disabled, and error states
+   * Enabled the styling of `menuItem`, `menuItemHighlight`, and `menuItemIcon` classes
+   * Renamed `muiIcon` class to `selectIcon`
+   * Fixed a bug when `disableSelectOnBlur` prop was passed
+   * Tweaked the behavior of `highlightFirstSuggestion()`
+   * Added support for `inputRef` prop
+ * MuiInput: Previously the field value was not updated until the user exited the input (`blur` event); now the value is updated as you type, enabling the form submit button sooner
+ * Various components: Changed the type of component props to `PropTypes.oneOfType([PropTypes.node, PropTypes.elementType])`
+ * Various components: Updated spacing to comply with linting rules
+ * ScrollTrigger: Refined functionality
+ * TooltipButton: Changed `buttonWrap` display to `inline-flex`
+ * Datatable: Added support for the `label` column prop where you can pass text, or a React element
+ * Component mixin: Updated list of fields in `cleanProps()`
+ * EndAdornment: Tweaked button spacing
+ * StartAdornment: Changed icon button to a `TooltipButton`
+ * Countries: Added `getRegionCode()` function
+ * FormSubmit: Tweaked button spacing
+ * ThemeStyles: Fixed minor bugs
+ * ModalTrigger: Now the trigger component is rendered using `instantiateComponent()`, the same as component props are rendered elsewhere
+ 
 1.16.0 / 2020-08-17
 ===================
 
@@ -40,6 +82,12 @@
    * Added default value for `paginationTerms`
    * New bonus component `DatatableFromArray` is a wrapper for `Datatable` that takes an array of objects and supports pagination
    
+1.13.2_2 / 2020-01-20
+=====================
+
+ * MuiSuggest: Removed `selectedOption` and `inputFormatted` from the component state
+ * TooltipIntl: Fixed bug: `titleValues` prop was not implemented
+
 1.13.2_1 / 2019-10-02
 =====================